+++ src/features/learn/constants.ts
@@ -19,3 +19,10 @@ export const PRELOAD_IMAGES = [
   '섬5.svg',
   '앉은-코코.svg',
 ];
+
+export const COLORS: Record<number, string> = {
+  1: '#007BFF', // 파란색
+  2: '#FAC608', // 노란색
+  3: '#F05650', // 빨간색
+  4: '#2DB400', // 초록색
+};
