+++ src/features/learn/ui/styles.ts
@@ -1,5 +1,7 @@
 import styled from 'styled-components';
 import { media } from '@style/media';
+import { getImageUrl } from '@utils/getImageUrl';
+import { animations } from '@style/animations';
 
 export const SectionBoxWrapper = styled.div`
   display: flex;
@@ -163,4 +165,190 @@ export const HandsUpCokoImg = styled.img`
   ${media.mobile} {
     display: none;
   }
+`;
+
+// 지도 윗 부분 이미지
+export const UpperBackgroundImg = styled.div`
+  width: 693px;
+  height: 147px;
+  background-image: url(${getImageUrl('배경1.webp')});
+  display: flex;
+  align-items: end;
+
+  ${media.mobile} {
+    display: none;
+  }
+`;
+
+// 버튼 섹션 전체를 감싸는 Wrapper (백그라운드 이미지 고정)
+export const EntireSectionContainer = styled.div<{
+  $isActiveBubble?: boolean;
+}>`
+  background-image: url(${getImageUrl('배경2.webp')});
+  padding-bottom: ${({ $isActiveBubble }) =>
+    $isActiveBubble ? '80px' : '0px'};
+  transition: margin-bottom 0.3s ease;
+
+  ${media.mobile} {
+    background-image: none;
+    padding-bottom: ${({ $isActiveBubble }) =>
+      $isActiveBubble ? '160px' : '60px'};
+  }
+`;
+
+// 섹션 하나를 감싸는 Wrapper
+export const SectionWrapper = styled.div`
+  padding-bottom: 20px;
+`;
+
+// 퀴즈 튜토리얼 링크 감싸는 Wrapper
+export const QuizTutorialLinkWrapper = styled.div`
+  display: flex;
+  justify-content: center;
+  margin-bottom: 20px;
+`;
+
+// 섹션 제목(이름)
+export const SectionTitle = styled.h2`
+  width: 693px;
+  font-size: 20px;
+  color: #ffffff;
+  text-align: center;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  ${media.mobile} {
+    color: #5f5f5f;
+    font-size: 20px;
+    max-width: 693px;
+    width: 100%;
+  }
+
+  &::before,
+  &::after {
+    content: '';
+    flex: 1;
+    height: 2px;
+    background: #ffffff;
+    margin: 0 55px;
+    max-width: 200px;
+
+    ${media.mobile} {
+      background: #5f5f5f;
+      margin: 0 35px;
+      max-width: 120px;
+    }
+  }
+`;
+
+// 버튼 배치를 위한 Wrapper (그리드 레이아웃)
+export const ButtonGrid = styled.section`
+  display: grid;
+  grid-template-columns: repeat(5, 1fr);
+  ${media.mobile} {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+`;
+
+// 앉은-코코 이미지
+export const SittingCoko = styled.img`
+  position: absolute;
+  top: -70px;
+  left: 50%;
+  transform: translateX(-50%);
+  z-index: 5;
+  width: 147px;
+  height: 120px;
+  animation: ${animations.fadeInAndDrop} 0.3s ease-out forwards;
+`;
+
+export const KeyboardButtonWrapper = styled.div`
+  position: relative;
+`;
+
+// 키캡(키보드 스위치) 버튼
+export const KeyboardButton = styled.button<{ $isLocked: boolean }>`
+  width: auto;
+  height: auto;
+  padding: 0;
+  border: none;
+  background-color: transparent;
+  margin-bottom: 20px;
+  cursor: ${({ $isLocked }) => ($isLocked ? 'not-allowed' : 'pointer')};
+
+  img {
+    width: 138.6px;
+    height: 72.68px;
+    display: block;
+    filter: ${({ $isLocked }) => ($isLocked ? 'grayscale(85%)' : 'none')};
+    opacity: ${({ $isLocked }) => ($isLocked ? 0.5 : 1)};
+  }
+`;
+
+// 파트 눌렀을 때 나오는 말풍선 (예시: 식별자)
+export const SpeechBubble = styled.div<{ $bgColor: string }>`
+  position: absolute;
+  background: ${({ $bgColor }) => $bgColor};
+  border-radius: 0.8em;
+  padding: 0.8em 1em;
+  width: 110px;
+  color: #ffffff;
+  text-align: center;
+  z-index: 5;
+  gap: 0.5em;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  left: 50%;
+  transform: translateX(-50%);
+
+  &:after {
+    content: '';
+    position: absolute;
+    bottom: 100%;
+    left: 50%;
+    width: 0;
+    height: 0;
+    border: 0.8em solid transparent;
+    border-bottom-color: ${({ $bgColor }) => $bgColor};
+    transform: translateX(-50%);
+  }
+
+  h3 {
+    font-size: 18px;
+    font-weight: 300;
+  }
+
+  ${media.mobile} {
+    width: 110px;
+    padding: 0.6em 0.8em;
+    top: calc(90%);
+    left: 50%;
+    right: auto;
+    transform: translateX(-50%);
+
+    &:after {
+      display: block;
+      top: -20px;
+      border-bottom-color: ${({ $bgColor }) => $bgColor};
+      border-top-color: transparent;
+    }
+
+    h3 {
+      font-size: 16px;
+    }
+  }
+`;
+
+// /quiz로 보내는 버튼
+export const GoToQuizButton = styled.button<{ $fontColor: string }>`
+  padding: 0.5em 1em;
+  border: none;
+  border-radius: 0.4em;
+  background-color: #ffffff;
+  color: ${({ $fontColor }) => $fontColor};
+  font-weight: 700;
 `;
